Move ConfigurationModal to its own folder + add logos
